requirements.txt
collections
numpy
tqdm
ipdb
lmdb
Pillow
scipy
matplotlib